use a buffered channel in the acceptQueue
There's a race condition here. A user might call Next to obtain the next
stream, and get a nil result. He then calls Chan get notified about new
incoming streams. If a new stream is accepted between the call to Next and the
call to Chan, he won't be notified about that stream, unless the channel is
buffered.